requirements.txt
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.12.0
anthropic>=0.18.0
aiosqlite>=0.19.0
structlog>=24.1.0
tenacity>=8.2.0
streamlit>=1.30.0
